+++ SPC_Coopenae/SPC_Coopenae.DATA/MetasIDP.cs
@@ -15,11 +15,11 @@ namespace SPC_Coopenae.DATA
     public partial class MetasIDP
     {
         public int IdMeta { get; set; }
-        public double Creditos { get; set; }
+        public decimal Creditos { get; set; }
         public int OtrosProductos { get; set; }
         public int TarjetasCredito { get; set; }
         public string Descripcion { get; set; }
         public int Estado { get; set; }
-        public double MinimoComisionarCDP { get; set; }
+        public decimal MinimoComisionarCDP { get; set; }
     }
 }
